requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
qrcode[pil]==7.4.2
gunicorn==22.0.0
python-dotenv==1.0.1
itsdangerous==2.2.0